+++ CHANGELOG.md
@@ -26,7 +26,10 @@
 * Rewrite `cyclomatic_complexity` rule using SwiftSyntax.  
   [Marcelo Fabri](https://github.com/marcelofabri)
 
-* Rewrite `redundant_void_return` rule using SwiftSyntax.  
+* Rewrite `redundant_void_return` rule using SwiftSyntax.
+  Also include redundant void return clauses for closures in addition to
+  functions. This can be disabled by configuring the rule with
+  `include_closures: false`.  
   [Marcelo Fabri](https://github.com/marcelofabri)
   [JP Simard](https://github.com/jpsim)